refactor: retire Big Five from active test pool

+++ app/schemas/psychologist_result.py
@@ -70,7 +70,7 @@ class ReviewCareerPatch(BaseModel):
     slug: str = Field(min_length=1)
     name: str = Field(min_length=1)
     holland_code: str
-    match_score: int
+    match_score: float
     description: str
     professions: list[str]
     skills_needed: list[str]
